Add possibility to use robust statistics
Since this was not abstracted away and means and standard deviations
were computed everywhere, we now return an object with "location" and
"scale" (and raw "data") and update the call sites everywhere. What
location and scale do is determined by a new parameter called
"parabolic_as_statistics" which can be either "standard" or "robust".