Create path joining util
Fixes #49. This hopefully works in most corner cases and avoids not
having to care about trailing / all over the place. In particular,
this is meant to solve the frequent problems where certain
configuration paths are required to have a trailing / for various
pieces of code to work, or, otherwise, to implement this logic
everywhere.