Do the work in a temporary folder
We move the temp folder to a final location after postfit is completed
to avoid considering partially completed runs as valid.

Currently the temporary folder is renamed to 'postfit' if there is no
error, and kept in the user system otherwise. Maybe should be refined
to delete the temp folder if a PostfitError is raised, but then we
need to see what to do with the log. It's probably fine as is.