Use ruamel_yaml
I recently updated reportengine to use ruamel_yaml instead of pyyaml.
It is better maintained and fixes some bugs I have seen in pyyaml
(such as incorrectly parsing `1.` as a string instead of a float).
The crucial advantage is that it provides a way to get error
messages with line numbers. It does have the disadvantage that its
packaging situation is crazy: The pypi package uses some unholy setup
that will slow down noticeably *all* your python scripts regardless of
whether thy actually use the package. The conda package fixes that but
the author of ruamel.yaml refused to merge the changes, and so the
conda package `ruamel_yaml` is different from the pypi package
`ruamel.yaml`. `reportengine.compat` abstracts away the difference.

Since this already works in reportengine, there is no reason it
shouldn't work in validphys.