Add functionality to select replicas for MC PDFs
This implements the filters discussed in:

<https://indico.cern.ch/event/647565/contributions/2707845/attachments/1519217/2372593/kassabov.pdf>

along with tools to visualize them. The interface could be better, in
that it could use validphys.core.Filter and  have a way to select the
filters that get applied in the runcard (which probably can be done
with a sufficiently clever production rule).

We also add a small class to plot the filtered replicas.

The x values are sampled differently which is why the number of
outliers is different than in the slides. I am thinking on removing
that filter altogether though.