Fix bug in chi2 bar plots
The ordering of the columns was not guaranteed to match the order of
the categories we infer (i.e. for the legend). The issue is, as way
too often, the handling of pandas MultiIndexes.  The 'levels' property
stores the unique values in each levels, but they are not guaranteed
to be ordered in any way. Order is recovered by using the 'labels'
property.

See:
https://stackoverflow.com/a/24489283/1007990

Instead we get the values for all the atomic columns and select the
unique items, which should correspond to fit names.