Add functionality to rewrite commondata file
This adds back the cuts as zeros entries, but still giving e.g. the
kinematics of the whole dataset. Zero entries are preferable to the
root data in the repository because they are more difficult to confuse
with e.g. fluctuations. Nans would have been better but it is more
complicated.

The new loading mechanism allows validphys2 to maintain its internal
consitency.