Fix sums of all nan slices
Hopefully this closes #16.

We restore the old behaviour whereby a sum of all_nans returns nan,
which we use in many places in the alpha_s code.

This was changed in an elephantine way (no deprecation period, no
warning) to return 0 instead, and an extra parameter was added to
allow to restore the old behaviour, but in a backwards incompatible
way, so we are forced to use the latest pandas version with this
change. I am fixing the version because I completely mistrust pandas
at this point.