Add interactive argument processing
Add command line options for the metadata arguments (as well as the
fits) and give the option to ask to complete them interactively when the
-i option is enabled.

An automatic script would avoid -i and pass all the arguments explicitly
instead.

Sensible defaults and completions are used when possible.

This still doesn't do anything with the options. It just provides a way
to enter them nicely.