Avoid exception when color cycle overflows
As it turns out, matplotlib's color specification is rather useless in
a programmatic setting, as we need to know how many colors we have to
use it properly (or use something like f'C{i%len(color_cycle)}') which
sort of defeats the purpose. Instead introduce a function that cycles
infinitely, but that will warn after the first cycle. Although ideally
we would make it a check that fires before computation is actually done.