Split up plots.py
plots.py became a lump of unrelated functionality that was getting very
difficult to understand and navigated. We split it in two:

 - pdfplots.py for things related to PDF plotting only.
 - dataplots.py for things involving data in some way.

Only one one-liner utility function, _scale_from_grid overlaps between
the two, and it has been moved to utils.py, used to guess the
appropriate scale (as in linear or log).

Also fix imports and remove some unneeded code.