Introduce CutsPolicy
This will allow to decide if we want to use internal cuts, fit-based
cuts or no cuts at all. To that end use_cuts evolves from a boolean to
a three state enum (I still need to decide if the enum provides enough
of an advantage in python).

The parsing preserves runcard-level backward
compatibility, by mapping the boolean to the corresponding enum option
(and printing a warning).

This will crash validphys at the moment as the actual loading still
needs to be done.