Add MatchedCuts
This allows to use `matched_datasets_from_dataspecs` to match cuts in
addition to datasets. Settings matched_cuts to true allows to  compare
datasets aligned exactly.

It is best for MatchedCuts to take the length of the dataset (which we
already know) and return an actual mask if all the other cuts turn out
to be None, so can it can interact in some way with other actions. We
already expect the cuts to be None, but it is madness to expect also
cuts.load() to be None.

We however add an optimization to not load from mask a dataset that is
already full.