Improve matched cuts interface
Create a separated production rule that produced the datasets with
matched cuts. This is much better in that it does away with the clumsy
check for the matched_cuts boolean (which only works if you had
matched_datasets_from_dataspecs in the first place) and in that it is
possible to use this production rule by itself, without having to use
mathced_datasets_from_dataspecs end the experiments scaffolding.