Add 'fromintersection' cuts policy
Add an option to compute cuts as the intersection of the result of
various ways to compute them. The idea is that if use_cuts is set to
fromintersection, the user is expected to provide a list of namespaces
called

cuts_intersection_spec

Then internal cuts are resolved for each namespace in the list, and the
final cuts are resolved as the intersection of all the results. Thus for
example one could do something like:

fit: NNPDF31_nnlo_as_0118_1000

theoryid: 53

use_cuts: fromintersection

datacuts:
    from_: fit

w2min:
    from_: datacuts

q2min:
    from_: datacuts

cuts_intersection_spec:
    - theoryid: 52
    - theoryid: 53

experiments:
    from_: fit

pdf:
    from_: fit

actions_:
    - experiments::experiment plot_fancy

Then each dataset will have the cuts corresponding to the intersection
between NLO and NNLO. We could as well have done something like:

cuts_intersection_spec:
    - theoryid: 52
      w2min: 10
    - theoryid: 53

with the obvious meaning. Variable resolution follows the usual
namespace chaining, so if a required key (e.g. q2min) is not found in
some of the namespaces in cut_intesection_spec, it will be looked up in
the upper namespace.

For this to be useful in fits, filters.py and vp-setupfit need to be
changed to use the existing cuts, rather than computing the internal
ones.