Use the internal cuts machinery in vp-setupfit
Do not use a separate code path for computing the masks to be written
and the internal cuts, but rather behave as if use_cuts: internal was
passed to validphys (which is how it is implemented now).

The main objective is to allow arbitrary cuts to be used by nnfit later
on.