Add pure Python FK parser
Add functionality to read fktables in pure python. This supports
transparently both compressed and plain tables. We have functionality to
process both types of sections defined in the spec, namely blobs and
option lists. We support types casting (and more general parsing) for
the option values and define various specialized parsing for various
blobs (at the moment based on numpy). These still need some work. At the
moment the code relies on the untested assumption that the FastKernel
section comes last. For the moment, return a dictionary with all the
fields.