Add ability to plot only lines in a band plot
This is useful for some of the plots in the th unc paper.

The interface needs to be improved in that we need a check similar to
check_pdf_normalize_to for the list of pdfs, so we can know that we
actually are plotting them and we support indexes (which is useful when
doing many comparisons).