Fix bug in different cuts
Fix  bug where we would get that the cuts are different if a mask with
all the data had been written for one of the fits and not for the other.
For example, this would report that the cuts are different:

```
use_cuts: fromfit
fits:
  - 190310-tg-nlo-DIS
  - 181126-si-nlo-central_DISonly

template_text: |
    {@print_different_cuts@}

actions_:
    - report(main=True)
```

even if they aren't. The reason why this bug existed in the first place
was the willingness to avoid loading datasets just to check the actual
number of datapoints, given that that is expensive. That has been worked
around since.