Fix errors related to missing processes
The process_lookup function is used in various places where we might
want to analize datasets that do not belong in the hardcoded per-process
dictionary. While that hardcoded dictionary is going away altogether,
(see #492), this needs to gracefully handle unknown datasets
immediately, as crashes can occur otherwise. This closes #497, which
showcases this problem.