Delete old function
The idea is not bad, but I don't believe this has been used that much
(or ever uploaded to the server), and we should have a modern version
that uses collect and log if we wanted to, which would be as easy as
rewriting it from scratch.

The current version is using some old version of results, and I can't be
bothered to fix it.