Add a with_cuts method to FKTableData
Allow to apply cuts to the new FKtable representation in a way that the
index is kept properly matched.

This is going to be useful to get full-fledged predictions out of this
object.

It may be useful to have this functionality at the level of the parser
when we get a nicer file format (in addition to or instead of here), but
for now it is probably fine to have it here.