Add various additional tools for predictions
Add central hadronic predictions and a component for linear
parametrizations to hadronic predictions. Specifically

  - Add an interface to compute central values only, when we don't
  need all the replicas.
  - Add an interface to compute a linearized approximation to hadronic
  predictions, suitable for tasks such as Hessian fits.

The interface is interesting in that it may open the door for TF (or
dask)  to use this code directly, since the only interface now is
through the grid_values function, which may well be something else.

Includes tests and documentation.