Add a plotting style with bigger labels
The default label sizes are intended for figures that occupy all of the
page, but we often want to have them at two or three columns. In that
case the plotting text is typically way too small. Add a style file that
makes them bigger.

At the moment one needs to run with

validphys --style <path/to>/biglabels.mplstyle

but I'll make a separate change to reportengine and vp to make that
simpler.