Bump version of libarchive
If I try to set up a clean environment on linux I get essentially this
error:

https://github.com/conda/conda/issues/8838

when trying to import NNPDF, which as said on the issue gets solved if
one installs libiconv manually. Indeed checking libarchive with ldd
I see that it is asking for libiconv and not finding it.

It seems there is some amount of dancing around of dancing around in the
conda forge channel with it being added

https://github.com/conda-forge/libarchive-feedstock/pull/36

and then removed

https://github.com/conda-forge/libarchive-feedstock/pull/59

and seems that that somehow can result in an incompatible environment.

Pull to the latest version (and be stricter with
pinning in hopes that resolves the problem.