Add stability condition table
Add a table to compute the stability of dataset covariance matrices.

There are a few things I am not convinced about, namely some amount of
code duplication, the fact that I am not sure which exact covmat
provider to use as input or that this should really be working with
square root matrices in the first place. It however solved the immediate
problem of having this in master.

Closes #1262.