Revert correlated shift functionality
This was broken on some code paths, seemed to incorrectly duplicate the
computation on others, and it is affected by very significative
conceptual limitations (see e.g.
https://github.com/NNPDF/nnpdf/pull/825#issuecomment-655331325) while it
is perhaps not so obvious what problem did it solve in the first place
(see e.g. https://github.com/NNPDF/nnpdf/issues/506#issuecomment-509213007).

It was not used in the 4.0 paper, and as a matter of empirical fact,
that did not causes the galaxy to be sucked into a black hole. The
coding related aspects could be improved upon, but that could benefit
from a fresh start. The algorithmic code lives in history and it is easy
enough to recover if needed.