Cache expensive dataframe operations outside loop
The make_replica function spends a substantial amount of time with
dataframe operations inside the relatively tight main loop. These are
repeat both for each non positive sample and for each replica, which
adds up to some very substantial runtime. With that generating a replica
on my laptop takes a few minutes in the worst case.

In particular the line

```
special_add_errors = pd.concat(special_add, axis=0, sort=True).fillna(0).to_numpy()
```

takes more than half a second by itself. This is possibly due to
a regression in pandas.

To improve that, add a new provider that computes all the dataframe
operations at once so they can be used for all replicas and all samples.