Fix callers of group_chi2_table
Commit 79604bd had introduced a level of indirection on the inputs of
groups_chi2_table to make sure that the data wasn't being reshuffled too
much. This however was not propagated to the actions that internally
made use of that functionality, of which there are three.

Fix the simpler one to use the original action with the right grouping
and add an additional collect over processes for the ones in the theory
covariance matrix module.