Index metadata in fits
Access the yaml inside the tar file and pull the metadata.

This is slow and inefficient, and simply lumps all the metadata in the
index file, although it should remain compatible with the old one.

It however really should be a different filename on the server probably.

Questions include how to make it fast and whether to pull all or some of
the metadata, and in which format.

Also probably to have some usable validphys UI.