Add ability to add filter rules
Sometimes we want to study something as a function of some cut. While now
it is possible to do, by copying all the filter rules, it is tedious.

A minimal solution is to allow for an extra key called
added_filter_rules, which, if present, should be a list behaving in the
same way as the internal cuts. The rules defined in there get parsed the
same way, but they key itself can be part of a namespace and therefore
vairiable in the same report.