remove use_theorycovmat key
This key was only used in a check, where it is forced to be set to true, though an error is raised if this is not the case, so it serves little purpose. Probably a remnant of a time where it could be used to toggle the use of the theory covmat.